Testing the First Level
So it’s Friday night and I’m having a few beers.  I put together a very
basic level one to get people excited here.  We need to think about the
main robot, because we have only first person here.  Should we have a
crosshair in the middle of the screen ( I think so)?  What do we want on
hud?  Health, ammo, and shields?   Are we going to have points?
Should we stop the player from dropping off the cliff into the target
practice range?  If we don’t we need to give him a path out.
For me I need to smooth the hills and find a better base texture.

Any ideas??